refactor: split FileExplorer into smaller components
- Created shared types in types.ts
- Split FileTreeNode into its own component
- Created FileDialog component
- Extracted logic into useFileExplorer hook
- Updated main FileExplorer to use new components
